feat: make hashing tf-idf vectorizer independent of ranker
