+++ deeppavlov/models/ranking/tfidf_ranker.py
@@ -18,7 +18,7 @@ import numpy as np
 
 from deeppavlov.core.common.registry import register
 from deeppavlov.core.common.log import get_logger
-from deeppavlov.core.models.estimator import Estimator
+from deeppavlov.core.models.estimator import Estimator, Component
 from deeppavlov.models.vectorizers.hashing_tfidf_vectorizer import HashingTfIdfVectorizer
 from deeppavlov.core.data.data_fitting_iterator import DataFittingIterator
 
@@ -39,55 +39,16 @@ class TfidfRanker(Estimator):
         top_n: a number of doc ids to return
         vectorizer: an instance of vectorizer class
         active: whether to return a number specified by :attr:`top_n` or all ids
-        tfidf_matrix: a loaded tfidf matrix
-        ngram_range: ngram range used when tfidf matrix was created
-        hash_size: hash size of the tfidf matrix
-        term_freqs: a dictionary with tfidf terms and their frequences
-        doc_index: a dictionary of doc ids and corresponding doc titles
         index2doc: inverted :attr:`doc_index`
         iterator: a dataset iterator used for generating batches while fitting the vectorizer
 
     """
-
-    def get_main_component(self) -> 'TfidfRanker':
-        """Temporary stub to run REST API
-
-        Returns:
-            self
-        """
-        return self
 
     def __init__(self, vectorizer: HashingTfIdfVectorizer, top_n=5, active: bool = True, **kwargs):
 
         self.top_n = top_n
         self.vectorizer = vectorizer
         self.active = active
-
-        if kwargs['mode'] != 'train':
-            if self.vectorizer.load_path.exists():
-                self.tfidf_matrix, opts = self.vectorizer.load()
-                self.ngram_range = opts['ngram_range']
-                self.hash_size = opts['hash_size']
-                self.term_freqs = opts['term_freqs'].squeeze()
-                self.doc_index = opts['doc_index']
-
-                self.vectorizer.doc_index = self.doc_index
-                self.vectorizer.term_freqs = self.term_freqs
-                self.vectorizer.hash_size = self.hash_size
-
-                self.index2doc = self.get_index2doc()
-            else:
-                self.iterator = None
-                logger.warning("TfidfRanker load_path doesn't exist, is waiting for training.")
-
-    def get_index2doc(self) -> Dict[Any, int]:
-        """Invert doc_index.
-
-        Returns:
-            inverted doc_index dict
-
-        """
-        return dict(zip(self.doc_index.values(), self.doc_index.keys()))
 
     def __call__(self, questions: List[str]) -> Tuple[List[Any], List[float]]:
         """Rank documents and return top n document titles with scores.
@@ -104,14 +65,14 @@ class TfidfRanker(Estimator):
         q_tfidfs = self.vectorizer(questions)
 
         for q_tfidf in q_tfidfs:
-            scores = q_tfidf * self.tfidf_matrix
+            scores = q_tfidf * self.vectorizer.tfidf_matrix
             scores = np.squeeze(
                 scores.toarray() + 0.0001)  # add a small value to eliminate zero scores
 
             if self.active:
                 thresh = self.top_n
             else:
-                thresh = len(self.doc_index)
+                thresh = len(self.vectorizer.doc_index)
 
             if thresh >= len(scores):
                 o = np.argpartition(-scores, len(scores) - 1)[0:thresh]
@@ -120,7 +81,7 @@ class TfidfRanker(Estimator):
             o_sort = o[np.argsort(-scores[o])]
 
             doc_scores = scores[o_sort]
-            doc_ids = [self.index2doc[i] for i in o_sort]
+            doc_ids = [self.vectorizer.index2doc[i] for i in o_sort]
             batch_doc_ids.append(doc_ids)
             batch_docs_scores.append(doc_scores)
